Add `Rational` class template
This class template models the mathematical concept of a rational
number, a number which can be represented as a fraction. It provides
common arithmetic operators so that it may be a drop-in replacement
for types like int or double which may be used in calculations.

The class makes an attempt to minimize the risk of overflow in
calculations as to maximize the amount of precision that can be
extracted from the underlying integer type. It maintains the
invariant that the fraction is always in its most reduced form.